Complete Page class refactor
